ref(compiler): move functions to translate.py
closes #1110
