+++ python/xorq/ibis_yaml/udf.py
@@ -110,21 +110,6 @@ def _scalar_udf_from_yaml(yaml_dict: dict, compiler: any) -> any:
     return node(*args).to_expr()
 
 
-@translate_to_yaml.register(bool)
-def _bool_to_yaml(value: bool, context: TranslationContext) -> dict:
-    return freeze(
-        {
-            "op": "bool",
-            "value": value,
-        }
-    )
-
-
-@register_from_yaml_handler("bool")
-def _bool_from_yaml(yaml_dict: dict, context: TranslationContext) -> bool:
-    return yaml_dict["value"]
-
-
 @translate_to_yaml.register(Schema)
 def _schema_to_yaml(schema: Schema, context: TranslationContext) -> dict:
     context.schema_registry.register_schema(schema)
@@ -147,32 +132,6 @@ def _schema_from_yaml(yaml_dict: dict, context: TranslationContext) -> Schema:
         {
             key: translate_from_yaml(value, context)
             for key, value in yaml_dict["value"].items()
-        }
-    )
-
-
-@translate_to_yaml.register(dt.DataType)
-def _datatype_to_yaml(dtype: dt.DataType, context: TranslationContext) -> dict:
-    return freeze(
-        {
-            "op": "DataType",
-            "type": type(dtype).__name__,
-        }
-        | {
-            argname: translate_to_yaml(arg, context)
-            for argname, arg in zip(dtype.argnames, dtype.args)
-        }
-    )
-
-
-@register_from_yaml_handler("DataType")
-def _datatype_from_yaml(yaml_dict: dict, context: TranslationContext) -> any:
-    typ = getattr(dt, yaml_dict["type"])
-    dct = toolz.dissoc(yaml_dict, "op", "type")
-    return typ(
-        **{
-            key: translate_from_yaml(value, context) if value is not None else None
-            for key, value in dct.items()
         }
     )
